Fixes for correct parsing and interpolation of DELFT and DORIS orbits
Difference wrt Getorb and other orbit parser on level of 1-3 mm's in
X and Y directions, and in Z direction 5-6 mm. The original
implementation would introduce errors on the level meters, that would
also scale with the size of the frame.

Issues resolved with working with 8th, instead of 3rd order Lagrange
polynomial, optimizations on the interpolation methods.

Changes summary:

 - Working with 8th instead of 3rd order polynomial.
 - Minor reorganizations in order of algo steps.
 - Optimization: precompute what can be precomputed.
 - Optimization: working with pseudo-coordinate systems for efficiency.
 - Also includes: clean-up of r&d code, refactorings, and comments.

Known problems:

 - Current implementation does not take into account if leapsecond is
   in the arc. This for sure effects interpolation of Delft (both ERS
   and Envisat), but I am not sure about Doris orbits.

Notes:

 - Not extensive testing had been performed with DORIS preliminary
   orbits.

 - BEAM and NEST do not share the same core, therefore Lagrange
   interpolation methods are at two places. BEAM changes will be
   submitted upstream.